+++ admins/pageflow/accounts.rb
@@ -14,51 +14,12 @@ module Pageflow
     form :partial => 'form'
 
     show :title => :name do |account|
-      attributes_table_for account do
-        row :name, :class => 'name'
-        row :default_file_rights, :class => 'default_file_rights'
-        row :created_at
-      end
-
-      attributes_table_for account.default_theming do
-        row :cname, :class => 'cname'
-        row :theme, :class => 'theme' do
-          account.default_theming.theme.name
-        end
-      end
+      render 'account_details', :account => account
+      render 'theming_details', :account => account
 
       div :class => 'columns' do
-        panel I18n.t('activerecord.models.user.other') do
-          if account.users.any?
-            table_for account.users, :i18n => User do
-              column :full_name do |user|
-                link_to user.full_name, admin_user_path(user)
-              end
-            end
-          else
-            div :class => "blank_slate_container" do
-              span :class => "blank_slate" do
-                I18n.t('admin.accounts.no_members')
-              end
-            end
-          end
-        end
-
-        panel t('activerecord.models.entry.other') do
-          if account.entries.any?
-            table_for account.entries, :i18n => Entry do
-              column :title do |entry|
-                link_to(entry.title, admin_entry_path(entry))
-              end
-            end
-          else
-            div :class => "blank_slate_container" do
-              span :class => "blank_slate" do
-                I18n.t('admin.accounts.no_entries')
-              end
-            end
-          end
-        end
+        render 'entries_panel', :account => account
+        render 'users_panel', :account => account
       end
     end